Basic UI integrated with selective webpage loading.
I have implemented the basic UI of the app, with a master page that
contians items that link to various parts of the app.  Webpages from the
CAM website can be easily converted into mobile views with
SelectiveWebPage, which allows you to specify the part of the page you
want to load without having the HTML hard-coded into the app.  So far,
Android support has been the main focus.