Improved performance and added UI features.
I improved performance by only modifying the SelectiveWebPage's
StackLayout's children instead of completely creating a new StackLayout
every time.  Also added a nice loading circle when the page is loading
instead of "Loading..."  Also created "MultiSelectiveWebPage," which
allows a SelectiveWebPage to have multiple mini subpages that can be
switched between with a button on the toolbar.