*Unstable* Haven't committed in days.
I did far too much to summarize what I did in however many characters
are provided by the summary box.  Heck, I don't even remember where the
last commit left off.  Well now I'm working on making the app fully
generated by the website.  Like, everything.  So I found the Google
Sheets document that controls the navigation bar and contains all the
links to all the webpages (exactly what I needed, yay).  So I wrote an
*insanely* complicated parser for it to generate a tree of nodes
containing the names of the sources along with the links to them.  So
the idea is I'll use that to generate the rest of the UI.  And I also
renamed "SelectiveWebPage" to "JunklessPage" because that seemed like a
more appopriate name.  And I'm using a more universal method to find the
meaty content of the webpage.  Anyways.  Sooooo many comments.  And like
60% of the existing .cs files won't exist after the new UI
implementation is gone.  And also there are soooo many comments.  Don't
mind them.  Will probably be fixed in the next commit.