The UI integration is almost done!
The UI is now COMPLETELY generated from the website!  Ahhhh so happy :)
Also added an "ExternalResourcesPage" which allows you to access links
on the website that lead to external sources like PDFs or other
websites.  And I removed all of the junk that was associated with the
old UI as well as all the old commented-out code.  And I documented
everything.  Probably the best commit yet!