Better splash screen, added search, improved UI.
The splash screen now has a loading circle and a label when it's
connecting to the Google Spreadsheet.  That way the user knows progress
is being made.  Also improved the efficiency of the master page items
and added better customization.  I made a custom Card class and CardView
class that the links and documents page now uses.  Most importantly, I
added a search feature, which allows the user to search for any
documents or pages in the app.  Pretty neato.