Added iOS support!!!
This app now supports iOS!!! Or does iOS support this app?...  idk
doesn't matter.  We did it!  Wasn't so hard.  Anyways.  The iOS version
was designed to look as close to the Android version as possible, while
still looking like an iPhone app.  All the colors and everything match
the Android version.  There are a few intricacies about the iOS
implementation that should be noted for future reference, though.  The
custom web view renderer works great, but iOS by default doesn't allow
HTTP requests from insecure servers, unfortunately including the CAM
website :'(  So you have to basically manually allow requests from the
CAM website in info.plist.  And the custom TabbedPage renderer.  Whew.
So in iOS tabbed pages can't scroll.  Instead, if there are too many
items to fit, there is a "more" button that you can click to see the
rest.  Ugly design imo, but that's just me.  But anyways, this also
means that iOS is willing to cram quite a few tab buttons on the bottom,
so the labels overlap if they're too long.  I spent quite a while trying
to figure out how to word wrap the bottom labels, which apparently is
not possible to do.  So instead, I figured I would make a custom
auto-ellipsis system.  While it's somewhat complex, it does in fact
shorten the name and add ellipses at the end if the name of the tab is
too long to fit in the space provided.  Works like a charm.  I also
updaded the logo to be the same as the splash screen icon.  It's cleaner
that way.  And fixed those random package warnings.  Fun stuff, good
times.  Project is looking good.