Fixed misc bugs, started major UI improvements.
A few miscellaneous bugs were fixed. The Google Sheets parser had a bug
in which it was possible to break it by entering a menu title without a
corresponding link. Also, the Xamarin version has been updated, which
fixed a Xamarin bug that prevented Android compatibility of versions 5.0
and lower. Also started on some major UI improvements, which (while
unfinished) will make the app generate a truly mobile-friendly webpage
that appears native, purely generated from the HTML on the website.